backend/requirement.txt
alembic==1.10.2
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
pydantic==1.10.7
PyJWT==2.6.0
python-decouple==3.7
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
starlette==0.22.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.20.0
websocket-client==1.5.1
websockets==10.4
win32-setctime==1.1.0
